Switch to setCanceledOnTouchOutside
http://developer.android.com/reference/android/app/Dialog.html#setCanceledOnTouchOutside%28boolean%29

Using setCancelable was causing an issue I noticed about 2 weeks ago when we started weeklies,
where the dialog could be cancelled by touching anywhere on the outside and so it could confuse
the user into thinking the download was cancelled.

While the download provider keeps the download going, users are users and sometimes they'll attempt
to download shit 2-3 times and possibly end up with a bad download and report false bugs to us.

Change-Id: Ia4a57344e72072533de4589efb441553576ac145