Ensure we load the section corresponding to what build the user is on
This just one of those things that just makes sense. If the user is opening
the app, they would probably like to see builds that are of the same. Now this
doesn't mean that we disallow them from viewing other sections but just let
loading up the section that corresponds with their current build.

- Added a null check to avoid startup crash on emulators

Change-Id: I2fdbc88efaf64c24bd9eb886c6e297f282c90999